MAINT: refactor flowproposal into a submodule
This should simplify implementing the MCMC version of FlowProposal
