+++ nessai/proposal/flowproposal/__init__.py
@@ -1,0 +1,5 @@
+from .flowproposal import FlowProposal
+
+__all__ = [
+    "FlowProposal",
+]
